--- src/main/CleanWayEvent.ts
@@ -1,3 +1,0 @@
-export enum CleanWayEvent {
-    TEST
-}
