--- src/main/CleanWayHandler.ts
@@ -1,5 +1,0 @@
-import CleanWaySubscriber from "./CleanWaySubscriber";
-
-export default interface CleanWayHandler {
-    accept(subscriber: CleanWaySubscriber): void;
-}
