--- src/main/CleanWaySubscriber.ts
@@ -1,3 +1,0 @@
-export default interface CleanWaySubscriber {
-    execute(): void;
-}
