+++ src/main/Test.ts
@@ -1,14 +1,14 @@
 import Annotation from "./Annotation";
-import CleanWayPubSub from "./CleanWayPubSub";
-import TestCase from "./TestCase";
-import {CleanWayEvent} from "./CleanWayEvent";
+import PubSub from "./pubsub/PubSub";
+import TestCase from "./testcase/TestCase";
+import {Event} from "./pubsub/Event";
 
 class Test implements Annotation<MethodDecorator> {
 
     public decorator(): MethodDecorator {
         return function (target: object, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
-            CleanWayPubSub.instance()
-                .subscribe(CleanWayEvent.TEST, new TestCase(target, propertyKey, descriptor));
+            PubSub.instance()
+                .subscribe(Event.TEST, new TestCase(target, propertyKey, descriptor));
         }
     }
 }
