+++ src/main/pubsub/Event.ts
@@ -1,0 +1,3 @@
+export enum Event {
+    TEST
+}
