+++ src/main/pubsub/Handler.ts
@@ -1,0 +1,5 @@
+import Subscription from "./Subscription";
+
+export default interface Handler {
+    accept(subscriber: Subscription): void;
+}
