+++ src/main/pubsub/Subscription.ts
@@ -1,0 +1,3 @@
+export default interface Subscription {
+    execute(): void;
+}
