+++ src/main/testcase/TestCaseRunner.ts
@@ -1,7 +1,7 @@
 import TestCase from "./TestCase";
-import CleanWayHandler from "./CleanWayHandler";
+import Handler from "../pubsub/Handler";
 
-export default class TestRunner implements CleanWayHandler {
+export default class TestCaseRunner implements Handler {
     public accept(testCase: TestCase) {
         testCase.execute();
     }
